+++ CHANGELOG.md
@@ -1,16 +1,29 @@
-# Change Log
+Change Log
+===
 
-## v1.1.0
+> Here's our record of all notable changes made to to this project
+
+v1.2.0
+---
+
+* Added new Hidden Items Pane Menu
+
+![pane-menu](https://explorer-exclude.s3.amazonaws.com/pane-menu.gif?v=1.2.0)
+
+v1.1.0
+---
 
 * Add Language Translations for [VS Code Supported Locales](https://code.visualstudio.com/docs/getstarted/locales#_available-locales)
 * Fix bug with Command Pallet showing Explorer Exclude as available commands when it should not
 
-## v1.0.1
+v1.0.1
+---
 
 * Extensions like Prophet Debugger were creating virtual folders in / for accessing log files, and this created a weird bug in the logic that tried to detect project paths
 * Added .github folder to help automate issue reporting
 * Updated logo with dark background for search results Marketplace page
 
-## v1.0.0
+v1.0.0
+---
 
 * Initial Release